Add support for LetsEncrypt
Well sort of... This adds an attribute which makes it easier to integrate this
with LE in other cookbooks. I decided not to actually implement LE directly in
the cookbook since the actual deployment of LE can vary widely depending on the
need. In the case for phpBB, this would not have worked since we're using this
cert for a webmail portal as well which might not be the case in other places.

However, I did test this with a test recipe using LE to ensure it actually works
properly!